fix: deserializing yaml documents fails with out of bounds
when a yaml Document is loaded it seems to contain a single
piece of Content, presumably representing the Document. assuming
this is a Mapping fails because a Mapping always has an even
number of pieces of Content (i.e. key/value pairs).

so when we find a Document with a single piece of Content
we dereference it so that it's child mappings get stored and
processed instead.